requirements.txt
python-telegram-bot~=22.1
pymongo~=4.13.0
python-dotenv==1.0.0
pydantic-settings~=2.9.1
aiohttp~=3.12.9